requirements.txt
chromadb>=0.4.0
sentence-transformers>=2.2.0
datasets>=2.14.0
python-dotenv>=1.0.0
groq>=0.4.0
httpx>=0.25.0
pytest>=7.4.0
rich>=13.0.0
